backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose==3.3.0
bcrypt==4.1.3
psycopg2-binary==2.9.9
email-validator==2.2.0
python-multipart==0.0.9
